pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "prompteval"
version = "0.1.0"
description = "Prompt evaluation and security auditing tool for LLMs"
requires-python = ">=3.10"
dependencies = [
    "click>=8.0",
    "pyyaml>=6.0",
    "jinja2>=3.0",
    "httpx>=0.25",
]

[project.optional-dependencies]
anthropic = ["anthropic>=0.30"]
openai = ["openai>=1.0"]
google = ["google-genai>=1.0"]
all = [
    "anthropic>=0.30",
    "openai>=1.0",
    "google-genai>=1.0",
]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
]

[project.scripts]
prompteval = "prompteval.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/prompteval"]
